validate update task

+++ src/main/java/controller/technicalmanager/TaskFormController.java
@@ -28,10 +28,9 @@ public class TaskFormController extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 		loadData(request, response);
-	
 
 	}
-	
+
 	private void loadData(HttpServletRequest request, HttpServletResponse response) {
 		try {
 			String idParam = request.getParameter("id");
@@ -39,26 +38,26 @@ public class TaskFormController extends HttpServlet {
 			Set<Integer> assignedStaffIds = null;
 			List<TaskDetail> taskDetail = new ArrayList<>();
 			if (idParam != null && !idParam.isEmpty()) {
-	            int taskId = Integer.parseInt(idParam);
-	            task = taskDao.getTaskById(taskId);
-	            taskDetail = taskDetailDao.getTaskDetail(taskId);
-	            assignedStaffIds = taskDao.getAssignedStaffIds(taskId);
-	        }
-			
+				int taskId = Integer.parseInt(idParam);
+				task = taskDao.getTaskById(taskId);
+				taskDetail = taskDetailDao.getTaskDetail(taskId);
+				assignedStaffIds = taskDao.getAssignedStaffIds(taskId);
+			}
+
 			List<CustomerIssue> issueList = issueDao.getAllIssues();
-		    List<User> staffList = userDao.getAllTechnicalStaff();
-			
+			List<User> staffList = userDao.getAllTechnicalStaff();
+
 			request.setAttribute("task", task);
-	        request.setAttribute("customerIssues", issueList);
-	        request.setAttribute("technicalStaffList", staffList);
-	        request.setAttribute("taskDetail", taskDetail);
-	        request.setAttribute("assignedStaffIds", assignedStaffIds);
+			request.setAttribute("customerIssues", issueList);
+			request.setAttribute("technicalStaffList", staffList);
+			request.setAttribute("taskDetail", taskDetail);
+			request.setAttribute("assignedStaffIds", assignedStaffIds);
 			request.getRequestDispatcher("view/admin/technicalmanager/taskForm.jsp").forward(request, response);
 		} catch (Exception e) {
 			System.out.print("Error");
 		}
 	}
-	
+
 	private void addNewTask(HttpServletRequest request, HttpServletResponse res) {
 		try {
 			String title = request.getParameter("title");
@@ -72,21 +71,21 @@ public class TaskFormController extends HttpServlet {
 			deadline = Timestamp.valueOf(deadlineStr + " 00:00:00");
 			int managerId = 2;
 			Timestamp now = new Timestamp(System.currentTimeMillis());
-			
-			if(title == null || title.trim().isEmpty()) {
+
+			if (title == null || title.trim().isEmpty()) {
 				request.setAttribute("errorTitle", "Không được để trống trường này");
 				loadData(request, res);
 				request.getRequestDispatcher("view/admin/technicalmanager/taskForm.jsp").forward(request, res);
 				return;
 			}
-			
-			if(deadline.before(now) || deadline.equals(now)) {
+
+			if (deadline.before(now) || deadline.equals(now)) {
 				request.setAttribute("errorDeadline", "Deadline phải hơn ngày hôm nay");
 				loadData(request, res);
 				request.getRequestDispatcher("view/admin/technicalmanager/taskForm.jsp").forward(request, res);
 				return;
 			}
-			
+
 			Task task = new Task();
 			task.setTitle(title);
 			task.setDescription(description);
@@ -94,19 +93,19 @@ public class TaskFormController extends HttpServlet {
 			task.setCustomerIssueId(customerIssueId);
 			int taskId = taskDao.addNewTask(task);
 			if (staffs != null && staffs.length != 0) {
-			    for (String staff : staffs) {
-			        int staffId = Integer.parseInt(staff);
-			        taskDetailDao.insertStaffToTask(taskId, staffId, deadline);
-			    }
+				for (String staff : staffs) {
+					int staffId = Integer.parseInt(staff);
+					taskDetailDao.insertStaffToTask(taskId, staffId, deadline);
+				}
 			}
 
-			
-			res.sendRedirect("task-list");;
+			res.sendRedirect("task-list");
+			;
 		} catch (Exception e) {
 			// TODO: handle exception
 		}
 	}
-	
+
 	private void updateTask(HttpServletRequest req, HttpServletResponse res) {
 		try {
 			int taskId = Integer.parseInt(req.getParameter("id"));
@@ -119,52 +118,62 @@ public class TaskFormController extends HttpServlet {
 			Timestamp deadline = null;
 			String deadlineStr = req.getParameter("deadline");
 			deadline = Timestamp.valueOf(deadlineStr + " 00:00:00");
-			
+
+			if (title == null || title.trim().isEmpty()) {
+				req.setAttribute("errorTitle", "Không được để trống trường này");
+				loadData(req, res);
+				req.getRequestDispatcher("view/admin/technicalmanager/taskForm.jsp").forward(req, res);
+				return;
+			}
+
 			taskDao.updateTask(taskId, title, description, managerId, customerIssueId);
-			
-			Set<Integer> newStaffSet = new HashSet<>();
-			
-			if(newStaffs != null) {
-				for (String staff : newStaffs) {
-					newStaffSet.add(Integer.parseInt(staff));
-				}
-			}
-			
+
 			Set<Integer> oldStaffSet = taskDao.getAssignedStaffIds(taskId);
-			
-			for (Integer staffId : newStaffSet) {
-				if(!oldStaffSet.contains(staffId)) {
-					taskDetailDao.insertStaffToTask(taskId, staffId, deadline);
-				}
-					
+
+			if (newStaffs != null && newStaffs.length > 0) {
+			    Set<Integer> newStaffSet = new HashSet<>();
+			    for (String staff : newStaffs) {
+			        newStaffSet.add(Integer.parseInt(staff));
+			    }
+
+			    for (Integer staffId : newStaffSet) {
+			        if (!oldStaffSet.contains(staffId)) {
+			            taskDetailDao.insertStaffToTask(taskId, staffId, deadline);
+			        }
+			    }
+
+			    for (Integer staffId : oldStaffSet) {
+			        if (!newStaffSet.contains(staffId)) {
+			            taskDetailDao.deleteStaffFromTask(taskId, staffId);
+			        }
+			    }
+
+			} else {
+			    for (Integer staffId : oldStaffSet) {
+			        taskDetailDao.deleteStaffFromTask(taskId, staffId);
+			    }
 			}
-			
-			for (Integer staffId : oldStaffSet) {
-				if(!newStaffSet.contains(staffId)) {
-					taskDetailDao.deleteStaffFromTask(taskId, staffId);
-				}
+
+			String deadlineStr2 = req.getParameter("deadline");
+			if (deadlineStr2 != null && !deadlineStr2.isEmpty()) {
+				Timestamp newDeadline = Timestamp.valueOf(deadlineStr2 + " 00:00:00");
+				taskDetailDao.updateDeadlineForTask(taskId, newDeadline);
 			}
-			 String deadlineStr2 = req.getParameter("deadline");
-		        if (deadlineStr2 != null && !deadlineStr2.isEmpty()) {
-		            Timestamp newDeadline = Timestamp.valueOf(deadlineStr2 + " 00:00:00");
-		            taskDetailDao.updateDeadlineForTask(taskId, newDeadline);
-		        }
-			
 			res.sendRedirect("task-list");
-			
+
 		} catch (Exception e) {
 			// TODO: handle exception
 		}
 	}
-	
+
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		String id = req.getParameter("id");
-	    if (id != null && !id.isEmpty()) {
-	        updateTask(req, resp);
-	    } else {
-	        addNewTask(req, resp);
-	    }
+		if (id != null && !id.isEmpty()) {
+			updateTask(req, resp);
+		} else {
+			addNewTask(req, resp);
+		}
 	}
 
 }
